cast to fix build

+++ coreapi/bellesip_sal/sal_impl.c
@@ -641,7 +641,7 @@ static int sal_add_listen_port(Sal *ctx, SalAddress* addr, bool_t is_tunneled){
 									sal_transport_to_string(sal_address_get_transport(addr)));
 	}
 	if (lp) {
-		belle_sip_listening_point_set_keep_alive(lp,ctx->keep_alive);
+		belle_sip_listening_point_set_keep_alive(lp,(int)ctx->keep_alive);
 		result = belle_sip_provider_add_listening_point(ctx->prov,lp);
 		if (sal_address_get_transport(addr)==SalTransportTLS) {
 			set_tls_properties(ctx);
@@ -716,7 +716,7 @@ void sal_set_keepalive_period(Sal *ctx,unsigned int value){
 	for (iterator=belle_sip_provider_get_listening_points(ctx->prov);iterator!=NULL;iterator=iterator->next) {
 		lp=(belle_sip_listening_point_t*)iterator->data;
 		if (ctx->use_tcp_tls_keep_alive || strcasecmp(belle_sip_listening_point_get_transport(lp),"udp")==0) {
-			belle_sip_listening_point_set_keep_alive(lp,ctx->keep_alive);
+			belle_sip_listening_point_set_keep_alive(lp,(int)ctx->keep_alive);
 		}
 	}
 }
@@ -1079,7 +1079,7 @@ int sal_generate_uuid(char *uuid, size_t len) {
 		return -1;
 	}
 	for (i = 0; i < 6; i++)
-		written+=snprintf(uuid+written,len-written,"%2.2x", uuid_struct.node[i]);
+		written+=snprintf(uuid+written,len-(size_t)written,"%2.2x", uuid_struct.node[i]);
 	uuid[len-1]='\0';
 	return 0;
 }
@@ -1106,7 +1106,7 @@ static void make_supported_header(Sal *sal){
 		const char *tag=(const char*)it->data;
 		size_t taglen=strlen(tag);
 		if (alltags==NULL || (written+taglen+1>=buflen)) alltags=reinterpret_cast<char *>(ms_realloc(alltags,(buflen=buflen*2)));
-		written+=snprintf(alltags+written,buflen-written,it->next ? "%s, " : "%s",tag);
+		written+=(size_t)snprintf(alltags+written,buflen-written,it->next ? "%s, " : "%s",tag);
 	}
 	if (alltags){
 		sal->supported=belle_sip_header_create("Supported",alltags);
@@ -1287,7 +1287,7 @@ unsigned char * sal_get_random_bytes(unsigned char *ret, size_t size){
 }
 
 char *sal_get_random_token(int size){
-	return belle_sip_random_token(reinterpret_cast<char *>(ms_malloc(size)),size);
+	return belle_sip_random_token(reinterpret_cast<char *>(ms_malloc((size_t)size)),(size_t)size);
 }
 
 unsigned int sal_get_random(void){
